API

API

API

Tables

Profiles

Formatting

Tables

SolverBenchmark.format_table — Function.

format_table(df, formatter, kwargs...)

Format the data frame into a table using formatter. Used by other table functions.

Inputs:

df::DataFrame: Dataframe of a solver. Each row is a problem.

formatter::Function: A function that formats its input according to its type. See LTXformat or MDformat for examples.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("%-10s", x))

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name").

Outputs:

header::Array{String,1}: header vector.

table::Array{String,2}: formatted table.

Base.join — Function.

df = join(stats, cols; kwargs...)

Join a dictionary of DataFrames given by stats. Column :id is required in all DataFrames. The resulting DataFrame will have column id and all columns cols for each solver.

Inputs:

stats::Dict{Symbol,DataFrame}: Dictionary of DataFrames per solver. Each key is a different solver;

cols::Array{Symbol}: Which columns of the DataFrames.

Keyword arguments:

invariant_cols::Array{Symbol,1}: Invariant columns to be added, i.e., columns that don't change depending on the solver (such as name of problem, number of variables, etc.);

hdr_override::Dict{Symbol,String}: Override header names.

Output:

df::DataFrame: Resulting dataframe.

SolverBenchmark.latex_table — Function.

latex_table(io, df, kwargs...)

Create a latex longtable using LaTeXTabulars of a dataframe of results, formatting the output for a publication-ready table.

Inputs:

io::IO: where to send the table, e.g.:

open("file.tex", "w") do io
  latex_table(io, df)
end

df::DataFrame: Dataframe of a solver. Each row is a problem.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("\textbf{%s}", x) |> safe_latex_AbstractString)`

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name"), where LaTeX escaping should be used if necessary.

We recommend using the safe_latex_foo functions when overriding formats, unless you're sure you don't need them.

SolverBenchmark.markdown_table — Function.

markdown_table(io, df, kwargs...)

Create a markdown table using PrettyTables of a dataframe of results, formatting the output.

Inputs:

io::IO: where to send the table, e.g.:

open("file.md", "w") do io
  markdown_table(io, df)
end

df::DataFrame: Dataframe of a solver. Each row is a problem.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("%-10s", x))

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name").

Profiles

BenchmarkProfiles.performance_profile — Function.

performance_profile(stats, cost)

Produce a performance profile comparing solvers in stats using the cost function.

Inputs:

stats::Dict{Symbol,DataFrame}: pairs of :solver => df;

cost::Function: cost function applyed to each df. Should return a vector with the cost of solving the problem at each row;

0 cost is not allowed;

If the solver did not solve the problem, return Inf or a negative number.

Examples of cost functions:

cost(df) = df.elapsed_time: Simple elapsed_time cost. Assumes the solver solved the problem.

cost(df) = (df.status .!= :first_order) * Inf + df.elapsed_time: Takes into consideration the status of the solver.

Formatting

SolverBenchmark.LTXformat — Function.

LTXformat(x)

Format x according to its type. For types Signed, AbstractFloat, AbstractString and Symbol, use a predefined formatting string passed to @sprintf and then the corresponding safe_latex_<type> function.

For type Missing, return "NA".

SolverBenchmark.MDformat — Function.

MDformat(x)

Format x according to its type. For types Signed, AbstractFloat, AbstractString and Symbol, use a predefined formatting string passed to @sprintf.

For type Missing, return "NA".

SolverBenchmark.safe_latex_AbstractFloat — Function.

safe_latex_AbstractFloat(s)

For floats. Bypasses Inf and NaN. Enclose both the mantissa and the exponent in \( and \).

SolverBenchmark.safe_latex_AbstractString — Function.

safe_latex_AbstractString(s)

For strings. Replaces _ by \_.

SolverBenchmark.safe_latex_Signed — Function.

safe_latex_Signed(s)

For signed integers. Encloses s in \( and \).

SolverBenchmark.safe_latex_Symbol — Function.

safe_latex_Symbol(s)

For symbols. Same as strings.