API

API

API

Tables

Formatting

Tables

SolverBenchmark.latex_table — Function.

latex_table(io, df, kwargs...)

Create a latex longtable using LaTeXTabulars of a dataframe of results, formatting the output for a publication-ready table.

Inputs:

io::IO: where to send the table, e.g.:

open("file.tex", "w") do io
  latex_table(io, df)
end

df::DataFrame: Dataframe of a solver. Each row is a problem.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("\textbf{%s}", x) |> safe_latex_AbstractString)`

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name"), where LaTeX escaping should be used if necessary.

We recommend using the safe_latex_foo functions when overriding formats, unless you're sure you don't need them.

Formatting

SolverBenchmark.LTXformat — Function.

LTXformat(x)

Format x according to its type. For types Signed, AbstractFloat, AbstractString and Symbol, use a predefined formatting string passed to @sprintf and then the corresponding safe_latex_<type> function.

For type Missing, return "NA".

SolverBenchmark.safe_latex_AbstractFloat — Function.

safe_latex_AbstractFloat(s)

For floats. Bypasses Inf and NaN. Enclose both the mantissa and the exponent in \( and \).

SolverBenchmark.safe_latex_AbstractString — Function.

safe_latex_AbstractString(s)

For strings. Replaces _ by \_.

SolverBenchmark.safe_latex_Signed — Function.

safe_latex_Signed(s)

For signed integers. Encloses s in \( and \).

SolverBenchmark.safe_latex_Symbol — Function.

safe_latex_Symbol(s)

For symbols. Same as strings.