API

API

Tables

PkgBenchmark

Profiles

Formatting

Tables

SolverBenchmark.format_table — Function

format_table(df, formatter, kwargs...)

Format the data frame into a table using formatter. Used by other table functions.

Inputs:

df::DataFrame: Dataframe of a solver. Each row is a problem.

formatter::Function: A function that formats its input according to its type. See LTXformat or MDformat for examples.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("%-10s", x))

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name").

Outputs:

header::Array{String,1}: header vector.

table::Array{String,2}: formatted table.

Base.join — Function

df = join(stats, cols; kwargs...)

Join a dictionary of DataFrames given by stats. Column :id is required in all DataFrames. The resulting DataFrame will have column id and all columns cols for each solver.

Inputs:

stats::Dict{Symbol,DataFrame}: Dictionary of DataFrames per solver. Each key is a different solver;

cols::Array{Symbol}: Which columns of the DataFrames.

Keyword arguments:

invariant_cols::Array{Symbol,1}: Invariant columns to be added, i.e., columns that don't change depending on the solver (such as name of problem, number of variables, etc.);

hdr_override::Dict{Symbol,String}: Override header names.

Output:

df::DataFrame: Resulting dataframe.

SolverBenchmark.latex_table — Function

latex_table(io, df, kwargs...)

Create a latex longtable of a DataFrame using LaTeXTabulars, and format the output for a publication-ready table.

Inputs:

io::IO: where to send the table, e.g.:

open("file.tex", "w") do io
  latex_table(io, df)
end

If left out, io defaults to stdout.

df::DataFrame: Dataframe of a solver. Each row is a problem.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("\textbf{%s}", x) |> safe_latex_AbstractString)`

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name"), where LaTeX escaping should be used if necessary.

We recommend using the safe_latex_foo functions when overriding formats, unless you're sure you don't need them.

SolverBenchmark.markdown_table — Function

markdown_table(io, df, kwargs...)

Create a markdown table from a DataFrame using PrettyTables and format the output.

Inputs:

io::IO: where to send the table, e.g.:

open("file.md", "w") do io
  markdown_table(io, df)
end

If left out, io defaults to stdout.

df::DataFrame: Dataframe of a solver. Each row is a problem.

Keyword arguments:

cols::Array{Symbol}: Which columns of the df. Defaults to using all columns;

ignore_missing_cols::Bool: If true, filters out the columns in cols that don't exist in the data frame. Useful when creating tables for solvers in a loop where one solver has a column the other doesn't. If false, throws BoundsError in that situation.

fmt_override::Dict{Symbol,Function}: Overrides format for a specific column, such as

fmt_override=Dict(:name => x->@sprintf("%-10s", x))

hdr_override::Dict{Symbol,String}: Overrides header names, such as hdr_override=Dict(:name => "Name").

PkgBenchmark

SolverBenchmark.bmark_results_to_dataframes — Function

stats = bmark_results_to_dataframes(results)

Convert PkgBenchmark results to a dictionary of DataFrames.

Inputs:

results::BenchmarkResults: the result of PkgBenchmark.benchmarkpkg()

Output:

stats::Dict{Symbol,DataFrame}: a dictionary of DataFrames containing the benchmark results per solver.

SolverBenchmark.judgement_results_to_dataframes — Function

stats = judgement_results_to_dataframes(judgement)

Convert BenchmarkJudgement results to a dictionary of DataFrames.

Inputs:

judgement::BenchmarkJudgement: the result of, e.g.,

commit = benchmarkpkg(mypkg)  # benchmark a commit or pull request
master = benchmarkpkg(mypkg, "master")  # baseline benchmark
judgement = judge(commit, master)

Output:

stats::Dict{Symbol,DataFrame}: a dictionary of DataFrames containing the target and baseline benchmark results.

SolverBenchmark.to_gist — Function

posted_gist = to_gist(results, p)

Create and post a gist with the benchmark results and performance profiles.

Inputs:

results::BenchmarkResults: the result of PkgBenchmark.benchmarkpkg()

p:: the result of profile_solvers().

Output:

the return value of GitHub.jl's create_gist().

posted_gist = to_gist(results)

Create and post a gist with the benchmark results and performance profiles.

Inputs:

results::BenchmarkResults: the result of PkgBenchmark.benchmarkpkg()

Output:

the return value of GitHub.jl's create_gist().

Profiles

BenchmarkProfiles.performance_profile — Function

performance_profile(stats, cost)

Produce a performance profile comparing solvers in stats using the cost function.

Inputs:

stats::Dict{Symbol,DataFrame}: pairs of :solver => df;

cost::Function: cost function applyed to each df. Should return a vector with the cost of solving the problem at each row;

0 cost is not allowed;

If the solver did not solve the problem, return Inf or a negative number.

Examples of cost functions:

cost(df) = df.elapsed_time: Simple elapsed_time cost. Assumes the solver solved the problem.

cost(df) = (df.status .!= :first_order) * Inf + df.elapsed_time: Takes into consideration the status of the solver.

SolverBenchmark.profile_solvers — Function

p = profile_solvers(stats, costs, costnames)

Produce performance profiles comparing solvers based on the data in stats.

Inputs:

stats::Dict{Symbol,DataFrame}: a dictionary of DataFrames containing the benchmark results per solver (e.g., produced by bmark_results_to_dataframes())

costs::Vector{Function}: a vector of functions specifying the measures to use in the profiles

costnames::Vector{String}: names to be used as titles of the profiles.

Keyword inputs:

width::Int: Width of each individual plot (Default: 400)

height::Int: Height of each individual plot (Default: 400)

Output: A Plots.jl plot representing a set of performance profiles comparing the solvers. The set contains performance profiles comparing all the solvers together on the measures given in costs. If there are more than two solvers, additional profiles are produced comparing the solvers two by two on each cost measure.

p = profile_solvers(results)

Produce performance profiles based on PkgBenchmark.benchmarkpkg results.

Inputs:

results::BenchmarkResults: the result of PkgBenchmark.benchmarkpkg().

SolverBenchmark.profile_package — Function

p = profile_package(judgement)

Produce performance profiles based on PkgBenchmark.BenchmarkJudgement results.

Inputs:

judgement::BenchmarkJudgement: the result of, e.g.,

commit = benchmarkpkg(mypkg)  # benchmark a commit or pull request
master = benchmarkpkg(mypkg, "master")  # baseline benchmark
judgement = judge(commit, master)

Formatting

SolverBenchmark.LTXformat — Function

LTXformat(x)

Format x according to its type. For types Signed, AbstractFloat, AbstractString and Symbol, use a predefined formatting string passed to @sprintf and then the corresponding safe_latex_<type> function.

For type Missing, return "NA".

SolverBenchmark.MDformat — Function

MDformat(x)

Format x according to its type. For types Signed, AbstractFloat, AbstractString and Symbol, use a predefined formatting string passed to @sprintf.

For type Missing, return "NA".

SolverBenchmark.safe_latex_AbstractFloat — Function

safe_latex_AbstractFloat(s)

For floats. Bypasses Inf and NaN. Enclose both the mantissa and the exponent in \( and \).

SolverBenchmark.safe_latex_AbstractString — Function

safe_latex_AbstractString(s)

For strings. Replaces _ with \_.

SolverBenchmark.safe_latex_Signed — Function

safe_latex_Signed(s)

For signed integers. Encloses s in \( and \).

SolverBenchmark.safe_latex_Symbol — Function

safe_latex_Symbol(s)

For symbols. Same as strings.